Fixed Key Animation
It was flickering and moving weirdly (it seemed to attach to the character but do nothing else; it was flickering and moving all over the sides of the screen weirdly without appearing to have been picked up). I made it into a parent and child with some similar components and some unique components to make it so the collider doesn't move and that fixed it for reasons I don't know.